+++ examples/kmm-sample/androidApp/src/main/java/io/realm/example/kmmsample/androidApp/MainActivity.kt
@@ -9,7 +9,6 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import io.realm.example.kmmsample.Calculator
 import io.realm.example.kmmsample.Greeting
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 
 fun greet(): String {
@@ -59,5 +58,8 @@ class MainActivity : AppCompatActivity() {
                 countTV.text = "History count: ${Calculator.history().size}"
             }
         }
+
+        // Smoke testing that we can access fields using `@PersistedName` after obfuscation
+        PersistedNameTest.testQueryByPersistedAndPublicName()
     }
 }