+++ packages/plugin-compiler/src/main/kotlin/io/realm/kotlin/compiler/Identifiers.kt
@@ -52,6 +52,7 @@ internal object Names {
     val REALM_ACCESSOR_HELPER_GET_UUID = Name.identifier("getUUID")
     val REALM_ACCESSOR_HELPER_GET_BYTE_ARRAY = Name.identifier("getByteArray")
     val REALM_ACCESSOR_HELPER_SET_VALUE = Name.identifier("setValue")
+    val REALM_ACCESSOR_HELPER_GET_REALM_ANY = Name.identifier("getRealmAny")
     val REALM_OBJECT_HELPER_GET_OBJECT = Name.identifier("getObject")
     val REALM_OBJECT_HELPER_SET_OBJECT = Name.identifier("setObject")
     val REALM_OBJECT_HELPER_SET_EMBEDDED_REALM_OBJECT = Name.identifier("setEmbeddedRealmObject")
@@ -105,6 +106,7 @@ internal object FqNames {
     val PRIMARY_KEY_ANNOTATION = FqName("io.realm.kotlin.types.annotations.PrimaryKey")
     val INDEX_ANNOTATION = FqName("io.realm.kotlin.types.annotations.Index")
     val IGNORE_ANNOTATION = FqName("io.realm.kotlin.types.annotations.Ignore")
+    val PERSISTED_NAME_ANNOTATION = FqName("io.realm.kotlin.types.annotations.PersistedName")
     val TRANSIENT_ANNOTATION = FqName("kotlin.jvm.Transient")
     val MODEL_OBJECT_ANNOTATION = FqName("io.realm.kotlin.internal.platform.ModelObject")
 
@@ -118,4 +120,5 @@ internal object FqNames {
     val KBSON_OBJECT_ID = FqName("org.mongodb.kbson.BsonObjectId")
     val REALM_UUID = FqName("io.realm.kotlin.types.RealmUUID")
     val REALM_MUTABLE_INTEGER = FqName("io.realm.kotlin.types.MutableRealmInt")
+    val REALM_ANY = FqName("io.realm.kotlin.types.RealmAny")
 }