+++ packages/plugin-compiler/src/main/kotlin/io/realm/kotlin/compiler/RealmModelSyntheticPropertiesGeneration.kt
@@ -28,6 +28,7 @@ import io.realm.kotlin.compiler.FqNames.OBJECT_REFERENCE_CLASS
 import io.realm.kotlin.compiler.FqNames.PRIMARY_KEY_ANNOTATION
 import io.realm.kotlin.compiler.FqNames.PROPERTY_INFO
 import io.realm.kotlin.compiler.FqNames.PROPERTY_TYPE
+import io.realm.kotlin.compiler.FqNames.REALM_ANY
 import io.realm.kotlin.compiler.FqNames.REALM_INSTANT
 import io.realm.kotlin.compiler.FqNames.REALM_MODEL_COMPANION
 import io.realm.kotlin.compiler.FqNames.REALM_OBJECT_ID
@@ -52,6 +53,7 @@ import io.realm.kotlin.compiler.Names.REALM_OBJECT_COMPANION_SCHEMA_METHOD
 import io.realm.kotlin.compiler.Names.SET
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.backend.common.lower.DeclarationIrBuilder
+import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSourceLocation
 import org.jetbrains.kotlin.descriptors.DescriptorVisibilities
 import org.jetbrains.kotlin.descriptors.Modality
 import org.jetbrains.kotlin.ir.ObsoleteDescriptorBasedAPI
@@ -139,6 +141,7 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
     private val realmObjectIdType: IrType = pluginContext.lookupClassOrThrow(REALM_OBJECT_ID).defaultType
     private val objectIdType: IrType = pluginContext.lookupClassOrThrow(KBSON_OBJECT_ID).defaultType
     private val realmUUIDType: IrType = pluginContext.lookupClassOrThrow(REALM_UUID).defaultType
+    private val realmAnyType: IrType = pluginContext.lookupClassOrThrow(REALM_ANY).defaultType
 
     private val kMutableProperty1Class: IrClass =
         pluginContext.lookupClassOrThrow(FqNames.KOTLIN_REFLECT_KMUTABLEPROPERTY1)
@@ -205,7 +208,8 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
             realmInstantType,
             objectIdType,
             realmObjectIdType,
-            realmUUIDType
+            realmUUIDType,
+            realmAnyType
         ).map { it.classifierOrFail }
     }
 
@@ -225,6 +229,19 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
         }
     }
 
+    /**
+     * Add RealmObjectCompanion.io_realm_kotlin_fields: Map<String, KProperty1<BaseRealmObject, Any?>>
+     *
+     * ```
+     * companion object : RealmObjectCompanion {                    // <--- TODO Double check
+     *      override val io_realm_kotlin_fields: Map<String, KProperty1<BaseRealmObject, Any?>> =
+     *          mapOf<String, KProperty1<BaseRealmObject, Any?>>(
+     *              Pair("propertyName", ClassName::property),      // same as: "propertyName" to ClassName::property
+     *              ...
+     *          )
+     * }
+     * ```
+     */
     @Suppress("LongMethod")
     fun addCompanionFields(
         clazz: IrClass,
@@ -258,7 +275,7 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
         ) { startOffset, endOffset ->
             IrConstImpl.string(startOffset, endOffset, pluginContext.irBuiltIns.stringType, className)
         }
-        // Add RealmObjectCompanion.io_realm_kotlin_fields: Map<String, KMutableProperty1<*, *>>
+        // Add io_realm_kotlin_fields
         companion.addValueProperty(
             pluginContext,
             realmObjectCompanionInterface,
@@ -283,10 +300,12 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                         UNDEFINED_OFFSET,
                         pluginContext.irBuiltIns.arrayClass.typeWith(companionFieldsElementType),
                         type,
+                        // Generate list of properties: List<Pair<String, KMutableProperty1<*, *>>>
                         properties!!.entries.map {
                             val property = it.value.declaration
                             val propertyType = if (it.value.isComputed) realmObjectPropertyType else
                                 realmObjectMutablePropertyType
+                            // Pair<String, KMutableProperty1<*, *>>()
                             IrConstructorCallImpl.fromSymbolOwner(
                                 startOffset = startOffset,
                                 endOffset = endOffset,
@@ -301,7 +320,7 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                         startOffset,
                                         endOffset,
                                         pluginContext.irBuiltIns.stringType,
-                                        property.name.identifier
+                                        it.value.persistedName
                                     )
                                 )
                                 putValueArgument(
@@ -386,6 +405,9 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
         val fields: MutableMap<String, SchemaProperty> =
             SchemaCollector.properties.getOrDefault(irClass, mutableMapOf())
 
+        // A map for tracking the property names and their source locations to ensure uniqueness
+        val persistedAndPublicNameToLocation = mutableMapOf<String, CompilerMessageSourceLocation>()
+
         val primaryKeyFields =
             fields.filter { it.value.declaration.backingField!!.hasAnnotation(PRIMARY_KEY_ANNOTATION) }
 
@@ -395,7 +417,7 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
         }
         val primaryKey: String? = when (primaryKeyFields.size) {
             0 -> null
-            1 -> primaryKeyFields.entries.first().key
+            1 -> primaryKeyFields.entries.first().value.persistedName
             else -> {
                 logError("RealmObject can only have one primary key", irClass.locationOf())
                 null
@@ -497,11 +519,24 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                         property.locationOf()
                                     )
                                 }
+                                val isIndexed = backingField.hasAnnotation(INDEX_ANNOTATION)
                                 if (isIndexed && backingField.type.classifierOrFail !in indexableTypes) {
                                     logError(
                                         "Indexed key ${property.name} is of type ${backingField.type.classifierOrFail.owner.symbol.descriptor.name} but must be of type ${indexableTypes.map { it.owner.symbol.descriptor.name }}",
                                         property.locationOf()
                                     )
+                                }
+
+                                val location = property.locationOf()
+                                val persistedName = value.persistedName
+                                val publicName = value.publicName
+
+                                // Ensure that the names are valid and do not conflict with prior persisted or public names
+                                ensureValidName(persistedName, persistedAndPublicNameToLocation, location)
+                                persistedAndPublicNameToLocation[persistedName] = location
+                                if (value.hasPersistedNameAnnotation) {
+                                    ensureValidName(publicName, persistedAndPublicNameToLocation, location)
+                                    persistedAndPublicNameToLocation[publicName] = location
                                 }
 
                                 IrCallImpl(
@@ -514,10 +549,10 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                 ).apply {
                                     dispatchReceiver = irGetObject(propertyClass.companionObject()!!.symbol)
                                     var arg = 0
-                                    // Name
-                                    putValueArgument(arg++, irString(entry.key))
+                                    // Persisted name
+                                    putValueArgument(arg++, irString(persistedName))
                                     // Public name
-                                    putValueArgument(arg++, irString(""))
+                                    putValueArgument(arg++, irString(publicName))
                                     // Type
                                     putValueArgument(
                                         arg++,
@@ -579,8 +614,8 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                     putValueArgument(
                                         arg++,
                                         if (type == linkingObjectType) {
-                                            val targetProperty = getLinkingObjectPropertyName(backingField)
-                                            irString(targetProperty.identifier)
+                                            val targetPropertyName = getLinkingObjectPropertyName(backingField)
+                                            irString(targetPropertyName)
                                         } else {
                                             irString("")
                                         }
@@ -752,4 +787,33 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
             endOffset,
             IrConstImpl.constFalse(startOffset, endOffset, pluginContext.irBuiltIns.booleanType)
         )
+
+    /**
+     * Ensure that persisted and public property names are unique.
+     *
+     * @param name the name to check uniqueness of
+     * @param existingNames the persisted and public names already parsed by the compiler
+     * @param location the location of the current property being parsed
+     */
+    private fun ensureValidName(name: String, existingNames: MutableMap<String, CompilerMessageSourceLocation>, location: CompilerMessageSourceLocation) {
+        if (name.isEmpty()) {
+            logError(
+                "Names must contain at least 1 character.",
+                location
+            )
+        }
+        if (existingNames.containsKey(name)) {
+            val duplicationLocation = existingNames[name]!!
+            if (location.line == duplicationLocation.line) {
+                // The message passed will only be contained in the compiler messages if `logDebug`
+                // or `logError` is used, not `logWarn` or `logInfo`. Thus, we opt for `logDebug` here.
+                logDebug("The Kotlin name and the persisted name are the same value: '$name'")
+            } else {
+                logError(
+                    "Kotlin names and persisted names must be unique. '$name' has already been used for the field on line ${duplicationLocation.line}.",
+                    location
+                )
+            }
+        }
+    }
 }